chore: move ChapterEndKeywords from repo to model package
