+++ backend/internal/service_new/validate.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/htchan/BookSpider/internal/model"
-	"github.com/htchan/BookSpider/internal/repo"
 )
 
 func isEnd(bk *model.Book) bool {
@@ -20,7 +19,7 @@ func isEnd(bk *model.Book) bool {
 	}
 
 	chapter := strings.ReplaceAll(bk.UpdateChapter, " ", "")
-	for _, keyword := range repo.ChapterEndKeywords {
+	for _, keyword := range model.ChapterEndKeywords {
 		if strings.Contains(chapter, keyword) {
 			return true
 		}
